+++ lib/asset/sql.go
@@ -25,6 +25,23 @@ func (p *Permissions) Scan(value interface{}) error {
 }
 
 // Value implements the driver.Valuer interface.
+// Simply returns the JSON-encoded representation of the Permission.
+func (p *Permission) Value() (driver.Value, error) {
+	return protojson.Marshal(p)
+}
+
+// Scan implements the sql.Scanner interface.
+// Simply decodes JSON into the Permission.
+func (p *Permission) Scan(value interface{}) error {
+	b, ok := value.([]byte)
+	if !ok {
+		return errors.NewError(errors.ErrByteArray, "cannot scan permission")
+	}
+
+	return protojson.Unmarshal(b, p)
+}
+
+// Value implements the driver.Valuer interface.
 // Simply returns the string representation of the AlgoCategory.
 func (c *AlgoCategory) Value() (driver.Value, error) {
 	return c.String(), nil
@@ -43,6 +60,52 @@ func (c *AlgoCategory) Scan(value interface{}) error {
 		return errors.NewError(errors.ErrInternal, "cannot scan algo category")
 	}
 	*c = AlgoCategory(v)
+
+	return nil
+}
+
+// Value implements the driver.Valuer interface.
+// Simply returns the string representation of the ComputeTaskStatus.
+func (ts *ComputeTaskStatus) Value() (driver.Value, error) {
+	return ts.String(), nil
+}
+
+// Scan implements the sql.Scanner interface.
+// Simply decodes a string into the ComputeTaskStatus.
+func (ts *ComputeTaskStatus) Scan(value interface{}) error {
+	s, ok := value.(string)
+	if !ok {
+		return errors.NewInternal("cannot scan task status: invalid string")
+	}
+
+	v, ok := ComputeTaskStatus_value[s]
+	if !ok {
+		return errors.NewInternal("cannot scan task status: unknown value")
+	}
+	*ts = ComputeTaskStatus(v)
+
+	return nil
+}
+
+// Value implements the driver.Valuer interface.
+// Simply returns the string representation of the ComputeTaskCategory.
+func (c *ComputeTaskCategory) Value() (driver.Value, error) {
+	return c.String(), nil
+}
+
+// Scan implements the sql.Scanner interface.
+// Simply decodes a string into the ComputeTaskCategory.
+func (c *ComputeTaskCategory) Scan(value interface{}) error {
+	s, ok := value.(string)
+	if !ok {
+		return errors.NewInternal("cannot scan task category: invalid string")
+	}
+
+	v, ok := ComputeTaskCategory_value[s]
+	if !ok {
+		return errors.NewInternal("cannot scan task category: unknown value")
+	}
+	*c = ComputeTaskCategory(v)
 
 	return nil
 }
@@ -96,23 +159,6 @@ func (d *DataManager) Scan(value interface{}) error {
 	}
 
 	return protojson.Unmarshal(b, d)
-}
-
-// Value implements the driver.Valuer interface.
-// Simply returns the JSON-encoded representation of the ComputeTask.
-func (a *ComputeTask) Value() (driver.Value, error) {
-	return protojson.Marshal(a)
-}
-
-// Scan implements the sql.Scanner interface.
-// Simply decodes JSON into the ComputeTask.
-func (a *ComputeTask) Scan(value interface{}) error {
-	b, ok := value.([]byte)
-	if !ok {
-		return errors.NewError(errors.ErrByteArray, "cannot scan compute task")
-	}
-
-	return protojson.Unmarshal(b, a)
 }
 
 // Value implements the driver.Valuer interface.