+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Remove compute task asset column (#619)
+
 ## [0.8.0] - 2022-04-11
 
 ### Added